requirements.txt
flask>=3.0
requests>=2.31
gspread>=6.0
pyobjc-framework-Vision>=10.0
pyobjc-framework-Cocoa>=10.0
python-dotenv>=1.0
